refactor: 리뷰 반영
- 서비스 패키지 내 항목을 도메인 패키지로 이동
- 메서드 명, 변수 명, 코드 가독성 개선
- 최소 반복 횟수를 상수로 추출
- 다중 검증 시 assertAll 사용
