+++ src/test/java/domain/LottoNumberGeneratorTest.java
@@ -1,7 +1,5 @@
-package service;
+package domain;
 
-import domain.LottoRule;
-import domain.Lottos;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -16,9 +14,9 @@ class LottoNumberGeneratorTest {
 
     @Test
     @DisplayName("요청된 개수만큼 로또를 생성해야 한다")
-    void 로또_개수만큼_생성되어야_한다() {
+    void 요청된_개수만큼_로또를_생성해야_한다() {
         // given
-        int repeatCount = 5;
+        int repeatCount = LottoNumberGenerator.MIN_REPEAT_COUNT + 1;
 
         // when
         Lottos lottos = lottoNumberGenerator.generateLottos(repeatCount);
@@ -32,7 +30,7 @@ class LottoNumberGeneratorTest {
     @DisplayName("각 로또는 정해진 개수를 가져야 한다")
     void 각_로또는_정해진_개수를_가져야_한다() {
         // given
-        int repeatCount = 5;
+        int repeatCount = LottoNumberGenerator.MIN_REPEAT_COUNT + 1;
 
         // when
         Lottos lottos = lottoNumberGenerator.generateLottos(repeatCount);
@@ -47,7 +45,7 @@ class LottoNumberGeneratorTest {
     @DisplayName("각 로또는 중복된 숫자를 가지면 안 된다")
     void 각_로또는_중복된_숫자를_가지면_안된다() {
         // given
-        int repeatCount = 10;
+        int repeatCount = LottoNumberGenerator.MIN_REPEAT_COUNT + 1;
 
         // when
         Lottos lottos = lottoNumberGenerator.generateLottos(repeatCount);
@@ -63,7 +61,7 @@ class LottoNumberGeneratorTest {
     @DisplayName("각 로또는 정해진 범위의 숫자만 포함해야 한다")
     void 각_로또는_정해진_범위의_숫자만_포함해야_한다() {
         // given
-        int repeatCount = 10;
+        int repeatCount = LottoNumberGenerator.MIN_REPEAT_COUNT + 1;
 
         // when
         Lottos lottos = lottoNumberGenerator.generateLottos(repeatCount);
@@ -81,11 +79,11 @@ class LottoNumberGeneratorTest {
     @DisplayName("잘못된 반복 횟수 입력 시 예외가 발생해야 한다")
     void 잘못된_반복_횟수_입력시_예외가_발생해야_한다() {
         // given
-        int invalidRepeatCount = 0;
+        int belowMinRepeatCount = LottoNumberGenerator.MIN_REPEAT_COUNT - 1;
 
         // when
         // then
-        assertThatThrownBy(() -> lottoNumberGenerator.generateLottos(invalidRepeatCount))
+        assertThatThrownBy(() -> lottoNumberGenerator.generateLottos(belowMinRepeatCount))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("회 이상 반복해야 합니다.");
     }
